refactor: 패키지 분리 (course 와 session)
