+++ src/main/java/nextstep/sessions/domain/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.sessions.domain;
 
 public enum SessionStatus {
     PREPARING {
