+++ src/main/java/nextstep/sessions/domain/Sessions.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.sessions.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
