+++ src/test/java/nextstep/sessions/domain/SessionTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.sessions.domain;
 
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUserTest;
