+++ src/test/java/nextstep/sessions/domain/SessionsTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.sessions.domain;
 
 
 import org.junit.jupiter.api.DisplayName;
